requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.4.2
boto3==1.29.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
qdrant-client==1.6.4
openai==1.3.0
pillow==10.1.0
numpy==1.26.2
pandas==2.1.3
pytest==7.4.3
httpx==0.25.1
redis==5.0.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-magic==0.4.27
aiofiles==23.2.1
tenacity==8.2.3
structlog==23.2.0
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0 